Test DeformationTestCase and SolidRotationTestCase
After adding velocity gradient tensor calculation in GeoMTK, the two
test cases are tested. The parcel shape is changed according to the flow
as expected, but there are cases where long and narrow shape is formed,
which I think is not good for accuracy. So next task is to handle these
cases by splitting long parcels.